:recycle: #50 refactor stores exports and naming
