+++ client/src/store/index.ts
@@ -1,0 +1,3 @@
+export { default as DebugStore } from './DebugStore'
+export { default as BenchStore } from './BenchStore'
+export { default as NavigationStore } from './NavigationStore'
